Enbale Both SSL and Plain TCP socket in single engine, you can set port use or not use ssl, set how many thread will be created for ssl trans, how many for plain trans, client use ssl can communicate with plain client as well

+++ ZoomPipeline_FuncSvr/smartlink/st_clientnode.h
@@ -51,7 +51,7 @@ protected:
 
     //UUID of this equipment
     bool m_bUUIDRecieved;
-    quint32 m_uuid;  //Client ID
+    quint64 m_uuid;  //Client ID
     //Client socket handle of this connection
     QObject * m_pClientSock;
